accommodation/requirements.txt
Django==4.1.7
pymongo==3.13.0
sqlparse==0.2.4
djongo==1.3.6
pytz
grpcio
grpcio-tools
